requirements.txt
# Financial Data Project Dependencies

# =============================================================================
# CORE / DATA
# =============================================================================
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL support
yfinance>=0.2.60

# =============================================================================
# WEB / DASHBOARDS
# =============================================================================
streamlit>=1.28.0
dash>=2.14.0
panel>=1.3.0
gradio>=4.0.0

# =============================================================================
# GRAFICOS / VISUALIZACION
# =============================================================================
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
bokeh>=3.3.0
altair>=5.2.0

# =============================================================================
# GRAFICOS FINANCIEROS
# =============================================================================
mplfinance>=0.12.10b
cufflinks>=0.17.3

# =============================================================================
# ANALISIS TECNICO
# =============================================================================
ta>=0.11.0
pandas-ta

# =============================================================================
# BACKTESTING
# =============================================================================
backtesting>=0.3.3
vectorbt>=0.26.0
backtrader>=1.9.78

# =============================================================================
# ML / PREDICCION
# =============================================================================
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
prophet>=1.1.0
statsmodels>=0.14.0

# =============================================================================
# METRICAS FINANCIERAS
# =============================================================================
quantstats>=0.0.62

# =============================================================================
# DATOS EXTERNOS
# =============================================================================
fredapi>=0.5.0
finnhub-python>=2.4.0
newsapi-python>=0.2.7
newspaper3k>=0.2.8  # Scraping de articulos de noticias

# =============================================================================
# IA / ASISTENTE
# =============================================================================
anthropic>=0.20.0
google-genai>=1.0.0
groq>=0.4.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0

# =============================================================================
# BASE DE DATOS VECTORIAL / RAG
# =============================================================================
chromadb>=0.4.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.0

# =============================================================================
# NLP / SENTIMENT ANALYSIS
# =============================================================================
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0

# pgvector for vector search (requires PostgreSQL extension)
pgvector>=0.2.0

# Redis for caching (optional, for scale)
redis>=5.0.0

# =============================================================================
# TABLAS / FORMATO
# =============================================================================
tabulate>=0.9.0
prettytable>=3.9.0
rich>=13.7.0

# =============================================================================
# REPORTES / PDF
# =============================================================================
reportlab>=4.0.0
fpdf2>=2.7.0
pdfkit>=1.0.0

# =============================================================================
# EXCEL / DOCUMENTOS
# =============================================================================
openpyxl>=3.1.0
xlsxwriter>=3.1.0
python-docx>=1.1.0

# =============================================================================
# IMAGENES / VISION
# =============================================================================
pillow>=10.0.0
opencv-python>=4.8.0
imageio>=2.33.0

# =============================================================================
# MAPAS / GEO
# =============================================================================
folium>=0.15.0
geopandas>=0.14.0

# =============================================================================
# ANIMACIONES
# =============================================================================
celluloid>=0.2.0

# =============================================================================
# HTTP / NETWORKING
# =============================================================================
httpx>=0.25.0
requests>=2.31.0

# =============================================================================
# CONFIGURACION
# =============================================================================
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# UTILIDADES
# =============================================================================
python-dateutil>=2.8.2
pytz>=2023.3

# =============================================================================
# SCHEDULER
# =============================================================================
APScheduler>=3.10.0

# =============================================================================
# DESARROLLO (opcional)
# =============================================================================
pytest>=7.4.0
black>=23.11.0
